Split src/lib/utils.ts by domain
Break the catch-all utils.ts into focused modules matching the existing $lib/utils/<domain>.ts convention: utils/json.ts (safeParse), utils/auth.ts (getBetterAuthErrorMessage), and utils/ui.ts (cn and the shadcn type helpers). App-level consumers of safeParse and getBetterAuthErrorMessage now import from the focused modules directly instead of the old catch-all path.

utils.ts itself becomes a thin compatibility re-export, since shadcn-svelte's generated components under components/ui/** (out of review scope) import cn and its companion types from the literal $lib/utils.js path as a CLI convention — updating ~100 generated files wasn't warranted for this refactor. Updated .fallowrc.json's dead-code ignore rules to match: utils.ts is now ignore-listed (it's only reachable from the ignored ui/** tree) and the WithElementRef/WithoutChildrenOrChild suppression now points at their new home in utils/ui.ts.

Fixes #265
