+++ src/lib/index.ts
@@ -1,2 +1,2 @@
-export { cn } from './utils';
+export { cn } from './utils/ui';
 export { logger } from './utils/logger';
