+++ src/lib/utils/json.ts
@@ -1,0 +1,8 @@
+export function safeParse<T>(json: string | null | undefined, fallback: T): T {
+	if (!json) return fallback;
+	try {
+		return JSON.parse(json) as T;
+	} catch {
+		return fallback;
+	}
+}
