+++ src/routes/(app)/journal/+page.server.ts
@@ -3,7 +3,7 @@ import { journalFilterSchema } from '$lib/schemas/journal';
 import { getUser } from '$lib/server/actions/auth-guard';
 import { getDb } from '$lib/server/db';
 import { journalEntries } from '$lib/server/db/schema';
-import { safeParse } from '$lib/utils';
+import { safeParse } from '$lib/utils/json';
 import { and, desc, eq, like } from 'drizzle-orm';
 
 import type { PageServerLoad } from './$types';
