+++ src/routes/(app)/journal/[id]/+page.server.ts
@@ -4,7 +4,7 @@ import { getUser, requireAuth } from '$lib/server/actions/auth-guard';
 import { getDb } from '$lib/server/db';
 import { journalEntries } from '$lib/server/db/schema';
 import { withAuditFieldsForUpdate } from '$lib/server/db/utils';
-import { safeParse } from '$lib/utils';
+import { safeParse } from '$lib/utils/json';
 import { error, fail, redirect } from '@sveltejs/kit';
 import { and, eq } from 'drizzle-orm';
 import { superValidate } from 'sveltekit-superforms';
