+++ src/routes/(app)/visits/+page.server.ts
@@ -3,8 +3,8 @@ import { getUser } from '$lib/server/actions/auth-guard';
 import { getDb } from '$lib/server/db';
 import { people, visits } from '$lib/server/db/schema';
 import { getVisitStatusThresholdsForUser } from '$lib/server/visit-status-settings';
-import { safeParse } from '$lib/utils';
 import { getTodayString } from '$lib/utils/date';
+import { safeParse } from '$lib/utils/json';
 import {
 	calculatePersonVisitStatus,
 	getStatusPriority,
